bot/requirements.txt
python-telegram-bot==21.9
httpx==0.27.2
python-dotenv==1.0.0
aiohttp==3.9.1
